examples/adk_fabric_agent/requirements.txt
google-adk @ git+https://github.com/google/adk-python.git
